Instr implementation is in progress

+++ src/vm/utils.h
@@ -1,6 +1,7 @@
 #ifndef __VM_UTILS_INCLUDED__
 #define __VM_UTILS_INCLUDED__
 
+#include <cstddef>
 #include <stdint.h>
 #include <ostream>
 #include <istream>
@@ -8,8 +9,8 @@
 namespace Ant {
   namespace VM {
 
-    int writeMultibyteInteger(uint64_t value, std::ostream &out);
-    int readMultibyteInteger(std::istream &in, uint64_t &value);
+    size_t writeMultibyteInteger(uint64_t value, std::ostream &out);
+    size_t readMultibyteInteger(std::istream &in, uint64_t &value);
 
   }
 }
